requirements.txt
# Web Framework
fastapi==0.135.2
uvicorn[standard]==0.42.0

# LLM & AI
langchain-core==1.2.23
langchain-google-genai==4.2.1
langgraph==1.1.3

# Vector Database
langchain-postgres==0.0.17
psycopg[binary]==3.3.3

# Utilities
python-dotenv==1.2.2
requests==2.33.0
pydantic==2.12.5
